Convert MAE to mlr task

Convert MAE to mlr task

Fun_MAE_to_taskFunc(MAE_obj, param.Y.name, param.covariates,
  param_positive_y_level, task_type)

Arguments

MAE_obj
MAE class

param.Y.name
Vector of dependent variable name

param.covariates
Vector of coaraiate variable(s) name

param_positive_y_level
if ClassifTask, value (character or numeric) to be considered as the positive factor outcome

Value

mlr's ClassifTask or RegrTask

Details

In case of individual MAE assay (omic) with multiple sub-assays, only first sub-assay will be used. MAE's helpers functions longFormat and wideFormat may not be best candidates, since mlr's functional data require complete subject structure for all 'assays' ('functionals'). this can be achieved by either removing non-complete subjects, or by creating dummy entities with NA. (which can later be imputed if required).

Either ClassifTask or RegrTask will be returned, based on the type of the param.Y.name variable

Examples

data(miniACC, package = 'MultiAssayExperiment') # ExpressionSet
miniACC

#> Loading required package: MultiAssayExperiment

#> 
#> Attaching package: 'MultiAssayExperiment'

#> The following object is masked _by_ '.GlobalEnv':
#> 
#>     miniACC

#> An object of class "MultiAssayExperiment"
#> Slot "ExperimentList":
#> ExperimentList class object of length 5: 
#>  [1] RNASeq2GeneNorm: SummarizedExperiment with 198 rows and 79 columns 
#>  [2] gistict: SummarizedExperiment with 198 rows and 90 columns 
#>  [3] RPPAArray: SummarizedExperiment with 33 rows and 46 columns 
#>  [4] Mutations: matrix with 97 rows and 90 columns 
#>  [5] miRNASeqGene: SummarizedExperiment with 471 rows and 80 columns 
#> 
#> Slot "colData":
#> DataFrame with 92 rows and 30 columns
#>                 patientID years_to_birth vital_status days_to_death
#>               <character>      <integer>    <integer>     <integer>
#> TCGA-OR-A5J1 TCGA-OR-A5J1             58            1          1355
#> TCGA-OR-A5J2 TCGA-OR-A5J2             44            1          1677
#> TCGA-OR-A5J3 TCGA-OR-A5J3             23            0            NA
#> TCGA-OR-A5J4 TCGA-OR-A5J4             23            1           423
#> TCGA-OR-A5J5 TCGA-OR-A5J5             30            1           365
#> ...                   ...            ...          ...           ...
#> TCGA-PK-A5H9 TCGA-PK-A5H9             27            0            NA
#> TCGA-PK-A5HA TCGA-PK-A5HA             63            0            NA
#> TCGA-PK-A5HB TCGA-PK-A5HB             63            0            NA
#> TCGA-PK-A5HC TCGA-PK-A5HC             44            0            NA
#> TCGA-P6-A5OG TCGA-P6-A5OG             45            1           383
#>              days_to_last_followup tumor_tissue_site pathologic_stage
#>                          <integer>       <character>      <character>
#> TCGA-OR-A5J1                    NA           adrenal         stage ii
#> TCGA-OR-A5J2                    NA           adrenal         stage iv
#> TCGA-OR-A5J3                  2091           adrenal        stage iii
#> TCGA-OR-A5J4                    NA           adrenal         stage iv
#> TCGA-OR-A5J5                    NA           adrenal        stage iii
#> ...                            ...               ...              ...
#> TCGA-PK-A5H9                   616           adrenal         stage ii
#> TCGA-PK-A5HA                  1201           adrenal          stage i
#> TCGA-PK-A5HB                  1293           adrenal               NA
#> TCGA-PK-A5HC                   679           adrenal        stage iii
#> TCGA-P6-A5OG                    NA           adrenal         stage iv
#>              pathology_T_stage pathology_N_stage      gender
#>                    <character>       <character> <character>
#> TCGA-OR-A5J1                t2                n0        male
#> TCGA-OR-A5J2                t3                n0      female
#> TCGA-OR-A5J3                t3                n0      female
#> TCGA-OR-A5J4                t3                n1      female
#> TCGA-OR-A5J5                t4                n0        male
#> ...                        ...               ...         ...
#> TCGA-PK-A5H9                t2                n0      female
#> TCGA-PK-A5HA                t1                n0        male
#> TCGA-PK-A5HB                NA                NA        male
#> TCGA-PK-A5HC                t4                n0      female
#> TCGA-P6-A5OG                t4                n0      female
#>              date_of_initial_pathologic_diagnosis radiation_therapy
#>                                         <integer>       <character>
#> TCGA-OR-A5J1                                 2000                no
#> TCGA-OR-A5J2                                 2004                no
#> TCGA-OR-A5J3                                 2008                no
#> TCGA-OR-A5J4                                 2000                no
#> TCGA-OR-A5J5                                 2000                no
#> ...                                           ...               ...
#> TCGA-PK-A5H9                                 2012                no
#> TCGA-PK-A5HA                                 2011                no
#> TCGA-PK-A5HB                                 2003               yes
#> TCGA-PK-A5HC                                 2011                no
#> TCGA-P6-A5OG                                 2011                no
#>                                 histological_type residual_tumor
#>                                       <character>    <character>
#> TCGA-OR-A5J1 adrenocortical carcinoma- usual type             r0
#> TCGA-OR-A5J2 adrenocortical carcinoma- usual type             r2
#> TCGA-OR-A5J3 adrenocortical carcinoma- usual type             r0
#> TCGA-OR-A5J4 adrenocortical carcinoma- usual type             r2
#> TCGA-OR-A5J5 adrenocortical carcinoma- usual type             r2
#> ...                                           ...            ...
#> TCGA-PK-A5H9 adrenocortical carcinoma- usual type             r0
#> TCGA-PK-A5HA adrenocortical carcinoma- usual type             r0
#> TCGA-PK-A5HB adrenocortical carcinoma- usual type             NA
#> TCGA-PK-A5HC adrenocortical carcinoma- usual type             r1
#> TCGA-P6-A5OG adrenocortical carcinoma- usual type             r2
#>              number_of_lymph_nodes        race              ethnicity
#>                          <integer> <character>            <character>
#> TCGA-OR-A5J1                    NA       white                     NA
#> TCGA-OR-A5J2                     0       white     hispanic or latino
#> TCGA-OR-A5J3                     0       white     hispanic or latino
#> TCGA-OR-A5J4                     2       white     hispanic or latino
#> TCGA-OR-A5J5                    NA       white     hispanic or latino
#> ...                            ...         ...                    ...
#> TCGA-PK-A5H9                    NA       asian not hispanic or latino
#> TCGA-PK-A5HA                     0          NA                     NA
#> TCGA-PK-A5HB                    NA          NA                     NA
#> TCGA-PK-A5HC                     0       asian not hispanic or latino
#> TCGA-P6-A5OG                     0       white not hispanic or latino
#>                Histology     C1A.C1B                              mRNA_K4
#>              <character> <character>                          <character>
#> TCGA-OR-A5J1  Usual Type         C1A steroid-phenotype-high+proliferation
#> TCGA-OR-A5J2  Usual Type         C1A steroid-phenotype-high+proliferation
#> TCGA-OR-A5J3  Usual Type         C1A               steroid-phenotype-high
#> TCGA-OR-A5J4  Usual Type          NA                                   NA
#> TCGA-OR-A5J5  Usual Type         C1A               steroid-phenotype-high
#> ...                  ...         ...                                  ...
#> TCGA-PK-A5H9  Usual Type         C1B                steroid-phenotype-low
#> TCGA-PK-A5HA  Usual Type         C1B                steroid-phenotype-low
#> TCGA-PK-A5HB  Usual Type         C1A               steroid-phenotype-high
#> TCGA-PK-A5HC  Usual Type          NA                                   NA
#> TCGA-P6-A5OG          NA          NA                                   NA
#>                     MethyLevel miRNA.cluster SCNA.cluster protein.cluster
#>                    <character>   <character>  <character>       <integer>
#> TCGA-OR-A5J1         CIMP-high       miRNA_1        Quiet              NA
#> TCGA-OR-A5J2          CIMP-low       miRNA_1        Noisy               1
#> TCGA-OR-A5J3 CIMP-intermediate       miRNA_6  Chromosomal               3
#> TCGA-OR-A5J4         CIMP-high       miRNA_6  Chromosomal              NA
#> TCGA-OR-A5J5 CIMP-intermediate       miRNA_2  Chromosomal              NA
#> ...                        ...           ...          ...             ...
#> TCGA-PK-A5H9          CIMP-low       miRNA_5        Quiet               3
#> TCGA-PK-A5HA         CIMP-high       miRNA_5  Chromosomal               2
#> TCGA-PK-A5HB         CIMP-high       miRNA_6        Noisy              NA
#> TCGA-PK-A5HC                NA            NA  Chromosomal              NA
#> TCGA-P6-A5OG                NA            NA           NA              NA
#>                      COC    OncoSign    purity    ploidy genome_doublings
#>              <character> <character> <numeric> <numeric>        <integer>
#> TCGA-OR-A5J1        COC3         CN2       0.9      1.95                0
#> TCGA-OR-A5J2        COC3    TP53/NF1      0.89      1.31                0
#> TCGA-OR-A5J3        COC2         CN2      0.93      1.25                0
#> TCGA-OR-A5J4          NA         CN1      0.87       2.6                1
#> TCGA-OR-A5J5        COC2    TP53/NF1      0.93      2.75                1
#> ...                  ...         ...       ...       ...              ...
#> TCGA-PK-A5H9        COC1    TP53/NF1      0.79         2                0
#> TCGA-PK-A5HA        COC1         CN2      0.83      1.69                0
#> TCGA-PK-A5HB        COC3    TP53/NF1      0.88      1.64                0
#> TCGA-PK-A5HC          NA    TP53/NF1      0.59      2.53                1
#> TCGA-P6-A5OG          NA          NA        NA        NA               NA
#>                    ADS
#>              <numeric>
#> TCGA-OR-A5J1     -0.08
#> TCGA-OR-A5J2     -0.84
#> TCGA-OR-A5J3      1.18
#> TCGA-OR-A5J4        NA
#> TCGA-OR-A5J5        -1
#> ...                ...
#> TCGA-PK-A5H9     -0.85
#> TCGA-PK-A5HA     -1.49
#> TCGA-PK-A5HB     -0.31
#> TCGA-PK-A5HC        NA
#> TCGA-P6-A5OG        NA
#> 
#> Slot "sampleMap":
#> DataFrame with 385 rows and 3 columns
#>               assay      primary                      colname
#>            <factor>  <character>                  <character>
#> 1   RNASeq2GeneNorm TCGA-OR-A5J1 TCGA-OR-A5J1-01A-11R-A29S-07
#> 2   RNASeq2GeneNorm TCGA-OR-A5J2 TCGA-OR-A5J2-01A-11R-A29S-07
#> 3   RNASeq2GeneNorm TCGA-OR-A5J3 TCGA-OR-A5J3-01A-11R-A29S-07
#> 4   RNASeq2GeneNorm TCGA-OR-A5J5 TCGA-OR-A5J5-01A-11R-A29S-07
#> 5   RNASeq2GeneNorm TCGA-OR-A5J6 TCGA-OR-A5J6-01A-31R-A29S-07
#> ...             ...          ...                          ...
#> 381    miRNASeqGene TCGA-PA-A5YG TCGA-PA-A5YG-01A-11R-A29W-13
#> 382    miRNASeqGene TCGA-PK-A5H8 TCGA-PK-A5H8-01A-11R-A29W-13
#> 383    miRNASeqGene TCGA-PK-A5H9 TCGA-PK-A5H9-01A-11R-A29W-13
#> 384    miRNASeqGene TCGA-PK-A5HA TCGA-PK-A5HA-01A-11R-A29W-13
#> 385    miRNASeqGene TCGA-PK-A5HB TCGA-PK-A5HB-01A-11R-A29W-13
#> 
#> Slot "metadata":
#> $title
#> [1] "Comprehensive Pan-Genomic Characterization of Adrenocortical Carcinoma"
#> 
#> $PMID
#> [1] "27165744"
#> 
#> $sourceURL
#> [1] "http://s3.amazonaws.com/multiassayexperiments/accMAEO.rds"
#> 
#> $RPPAfeatureDataURL
#> [1] "http://genomeportal.stanford.edu/pan-tcga/show_target_selection_file?filename=Allprotein.txt"
#> 
#> $colDataExtrasURL
#> [1] "http://www.cell.com/cms/attachment/2062093088/2063584534/mmc3.xlsx"
#> 
#> 
#> Slot "drops":
#> list()
#> 

Fun_MAE_to_taskFunc(miniACC, param.Y.name = 'vital_status', param.covariates = c('gender','days_to_death'), param_positive_y_level = '1')

#> Error in MAE_obj %>% colData %>% data.frame %>% rownames_to_column("primary"): could not find function "%>%"